commit upload.controller

+++ controllers/project.controller.js
@@ -1,7 +1,4 @@
-const { isValidObjectId } = require("mongoose");
 const ProjectModel = require("../models/project.model");
-const UserModel = require("../models/user.model");
-const { addProjectError } = require("../errors/project.errors");
 const CryptoJS = require("crypto-js");
 
 const encrypt = (text, token) => {
@@ -28,6 +25,7 @@ const addCom = async (idProject, idUser, text) => {
     { upsert: true, new: true }
   )
     .select("-connectionDescription")
+    .populate("comments.commenterId", "firstname name -_id")
     .exec()
     .then((docs) => {
       return { res: docs };
@@ -38,53 +36,55 @@ const addCom = async (idProject, idUser, text) => {
 };
 
 module.exports.addProject = async (req, res) => {
-  const {
-    customerName,
-    numElecDraw,
-    numCommand,
-    numMachine,
-    sector,
-    designation,
-    comment,
-    creatorId,
-    connectionMethod,
-    connectionAccount,
-    connectionLogin,
-    connectionPassword,
-  } = req.body;
-  const project = await ProjectModel.create({
-    customerName,
-    numElecDraw,
-    numCommand,
-    numMachine,
-    machineDescription: {
+  if (res.statusCode !== 401) {
+    const {
+      customerName,
+      numElecDraw,
+      numCommand,
+      numMachine,
       sector,
       designation,
       comment,
-    },
-    connectionDescription: {
+      creatorId,
       connectionMethod,
       connectionAccount,
       connectionLogin,
-      connectionPassword: connectionPassword
-        ? encrypt(connectionPassword, process.env.TOKEN_CRYPT)
-        : "",
-    },
-  })
-    .then(async (docs) => {
-      const response = await addCom(
-        docs._id,
-        creatorId,
-        "Creation of the project"
-      );
-      if (response.res) res.status(201).json(response.res);
-      else res.status(400).json(response.err);
+      connectionPassword,
+    } = req.body;
+    const project = await ProjectModel.create({
+      customerName,
+      numElecDraw,
+      numCommand,
+      numMachine,
+      machineDescription: {
+        sector,
+        designation,
+        comment,
+      },
+      connectionDescription: {
+        connectionMethod,
+        connectionAccount,
+        connectionLogin,
+        connectionPassword: connectionPassword
+          ? encrypt(connectionPassword, process.env.TOKEN_CRYPT)
+          : "",
+      },
     })
-    .catch((err) => res.status(400).json("Error : addProject function"));
+      .then(async (docs) => {
+        const response = await addCom(
+          docs._id,
+          creatorId,
+          "Creation of the project"
+        );
+        if (response.res) res.status(201).json(response.res);
+        else res.status(400).json(response.err);
+      })
+      .catch((err) => res.status(400).json("Error : addProject function"));
+  }
 };
 
 module.exports.updateProject = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     const {
       customerName,
       numElecDraw,
@@ -135,7 +135,7 @@ module.exports.updateProject = async (req, res) => {
 };
 
 module.exports.removeProject = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     return await ProjectModel.findByIdAndRemove(req.params.id)
       .then((docs) => {
         return res.status(200).json("The project has been removed");
@@ -145,65 +145,64 @@ module.exports.removeProject = async (req, res) => {
 };
 
 module.exports.getProject = async (req, res) => {
-  await ProjectModel.findById(req.params.id, (err, docs) => {
-    if (err || !docs) {
-      // Control if the id of the project exists
-      return res.status(400).json({ res: "Project unknown" });
-    } else {
-      return ProjectModel.findById(req.params.id, (err, docs) => {
-        if (!err) return res.status(200).json({ res: docs });
-        else return res.status(201).json({ res: err });
-      });
-    }
-  });
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
+    return await ProjectModel.findById(req.params.id)
+      .select("-connectionDescription")
+      .populate("comments.commenterId", "firstname name -_id")
+      .exec()
+      .then((docs) => res.status(200).json(docs))
+      .catch((err) => res.status(400).json("Error : getProject function"));
+  }
 };
 
-module.exports.getAllProjectLight = async (req, res) => {
-  const projects = await ProjectModel.find()
-    .select("-machineDescription")
-    .select("-connectionDescription")
-    .select("-comments");
-  /*
-    .select("-connectionDescription.connectionAccount")
-    .select("-connectionDescription.connectionLogin")
-    .select("-connectionDescription.connectionPassword");*/
-  return res.status(200).json({ res: projects });
-};
-
-module.exports.getAllProjectFull = async (req, res) => {
-  const projects = await ProjectModel.find()
-    .select("-connectionDescription.connectionAccount")
-    .select("-connectionDescription.connectionLogin")
-    .select("-connectionDescription.connectionPassword");
-  return res.status(200).json({ res: projects });
+module.exports.getAllProject = async (req, res) => {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
+    await ProjectModel.find()
+      .select("-connectionDescription")
+      .populate("comments.commenterId", "firstname name -_id")
+      .exec()
+      .then((docs) => {
+        if (req.body.filter === "full") res.status(200).json(docs);
+        else {
+          const responses = [];
+          docs.forEach((element) => {
+            responses[responses.length] = {
+              customerName: element.customerName,
+              numCommand: element.numCommand,
+              numMachine: element.numMachine,
+              numElecDraw: element.numElecDraw,
+            };
+          });
+          res.status(200).json(responses);
+        }
+      })
+      .catch((err) => res.status(400).json("Error : getAllProject function"));
+  }
 };
 
 module.exports.getPasswordConnection = async (req, res) => {
-  await ProjectModel.findById(req.params.id, (err, docs) => {
-    if (err || !docs) {
-      // Control if the id of the project exists
-      return res.status(400).json({ res: "Project unknown" });
-    } else {
-      return ProjectModel.findById(req.params.id, (err, docs) => {
-        if (!err) {
-          return res.status(200).json({
-            res: {
-              account: docs.connectionDescription.connectionAccount,
-              login: docs.connectionDescription.connectionLogin,
-              password: decrypt(
-                docs.connectionDescription.connectionPassword,
-                process.env.TOKEN_CRYPT
-              ),
-            },
-          });
-        } else return res.status(201).json({ res: err });
-      });
-    }
-  });
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
+    return await ProjectModel.findById(req.params.id)
+      .then((docs) => {
+        const response = {
+          method: docs.connectionDescription.connectionMethod,
+          account: docs.connectionDescription.connectionAccount,
+          login: docs.connectionDescription.connectionLogin,
+          password: decrypt(
+            docs.connectionDescription.connectionPassword,
+            process.env.TOKEN_CRYPT
+          ),
+        };
+        return res.status(200).json(response);
+      })
+      .catch((err) =>
+        res.status(400).json("Error : getPasswordConnection function")
+      );
+  }
 };
 
 module.exports.addComment = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     const response = await addCom(
       req.params.id,
       req.body.userId,
@@ -215,7 +214,7 @@ module.exports.addComment = async (req, res) => {
 };
 
 module.exports.updateComment = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     await ProjectModel.findById(req.params.id)
       .select("comments")
       .populate("comments.commenterId", "firstname name -_id")
@@ -236,7 +235,7 @@ module.exports.updateComment = async (req, res) => {
 };
 
 module.exports.removeComment = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     await ProjectModel.findByIdAndUpdate(
       req.params.id,
       {
@@ -256,7 +255,7 @@ module.exports.removeComment = async (req, res) => {
 };
 
 module.exports.getComment = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     await ProjectModel.findById(req.params.id)
       .select("comments")
       .populate("comments.commenterId", "firstname name -_id")
@@ -269,7 +268,7 @@ module.exports.getComment = async (req, res) => {
 };
 
 module.exports.getAllComment = async (req, res) => {
-  if (res.statusCode !== 400) {
+  if (res.statusCode !== 400 && res.statusCode !== 401) {
     await ProjectModel.findById(req.params.id)
       .select("comments")
       .populate("comments.commenterId", "firstname name -_id")
